Fix: Update imports after supabase→database rename and clean up stale doc refs
