+++ .gitignore
@@ -46,3 +46,4 @@ CLAUDE.md
 
 # Factory/Agent tooling
 .factory/
+CONTINUITY.md
