+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const a=(a,c)=>(a=new URL(a+".js",c).href,s[a]||new Promise(s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()}).then(()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e}));self.define=(c,i)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(s[n])return;let t={};const f=e=>a(e,n),r={module:{uri:n},exports:t,require:f};s[n]=Promise.all(c.map(e=>r[e]||f(e))).then(e=>(i(...e),t))}}define(["./workbox-f52fd911"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"6e2cf7216f3de5149e7a89f90c739dac"},{url:"/_next/static/AGGcf33yqn87hELktIsI3/_buildManifest.js",revision:"2ec694eb52ae4f523f265a46bae4d768"},{url:"/_next/static/AGGcf33yqn87hELktIsI3/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/23-e49c25ff4a2d28a9.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/294-2fa965b3689286af.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/3-2f6302d2d82ccbf4.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/51-3dd90a3c681c6cc8.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/638-fd1a0257b54ad914.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/696.643551c3285e05af.js",revision:"643551c3285e05af"},{url:"/_next/static/chunks/89-2472bce46db7878c.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/aaea2bcf-751e554790cca40d.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/app/_not-found/page-93d783c8c64ea450.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/app/dashboard/page-6beeff44e7a37c24.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/app/demo/page-93e84db19a0bdb4b.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/app/layout-d6a93d3cfd64ad7c.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/app/page-e6bdc8a24b6d5cd8.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/d0deef33.afff9b3bf7059c8b.js",revision:"afff9b3bf7059c8b"},{url:"/_next/static/chunks/fd9d1056-a7223fb5cd24b6ec.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/framework-f66176bb897dc684.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/main-aa26475113dfe328.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/main-app-28eecd8be4b1a924.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/pages/_app-6a626577ffa902a4.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/pages/_error-1be831200e60c5c0.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-b12133d75871ed5d.js",revision:"AGGcf33yqn87hELktIsI3"},{url:"/_next/static/css/857aa62baa17c894.css",revision:"857aa62baa17c894"},{url:"/_next/static/css/ab0f62cb2562f093.css",revision:"ab0f62cb2562f093"},{url:"/_next/static/media/26d4368bf94c0ec4-s.p.woff2",revision:"d28498895ed7050c37c94ddd026a4670"},{url:"/_next/static/media/2801417b65625cf5-s.woff2",revision:"d313794bd9fda8b0ad44189b27348eab"},{url:"/_next/static/media/28793f5c5e3d822d-s.woff2",revision:"c352fb162d499702d53f0dbfaab7b58f"},{url:"/_next/static/media/558ca1a6aa3cb55e-s.p.woff2",revision:"570751c5f8b418972c1976160ba6ed85"},{url:"/_next/static/media/64d784ea54a4acde-s.woff2",revision:"8a5b33d747f0cfaac631ad00bd5bcba2"},{url:"/_next/static/media/6c25f6e897d845a3-s.p.woff2",revision:"1a662f8f0b096a4b0954f9a6017ad87d"},{url:"/_next/static/media/6d831b18ae5b01dc-s.woff2",revision:"e6155c5cfacf3867c500daf0ebcba222"},{url:"/_next/static/media/7b19b489dc6743ba-s.woff2",revision:"fa052422282eb26cff400131e5af4da1"},{url:"/_next/static/media/8b44c7e6549520b2-s.woff2",revision:"7113097e6195b2704f2923f6989dd992"},{url:"/_next/static/media/ac0e76ddaeeb7981-s.woff2",revision:"6465b62dd12646a816e0d80f024ab07f"},{url:"/_next/static/media/b3bf17a9041d9433-s.woff2",revision:"ded54b0f07a0919cf7e27d7880a56e0a"},{url:"/_next/static/media/c9c3823090ec8b55-s.woff2",revision:"625be018001610bec1a5c0ba45d9988c"},{url:"/_next/static/media/edc640959b0c7826-s.woff2",revision:"5508edf7c10fe677025b8c88a2578acb"},{url:"/_next/static/media/ff71da380fbe67dd-s.woff2",revision:"60d32697500d4779da3725134067ad31"},{url:"/_next/static/media/layers-2x.9859cd12.png",revision:"9859cd12"},{url:"/_next/static/media/layers.ef6db872.png",revision:"ef6db872"},{url:"/_next/static/media/marker-icon.d577052a.png",revision:"d577052a"},{url:"/icon.svg",revision:"4a4a6e337480abfb479c94e37c5f214d"},{url:"/icons/icon-180.png",revision:"4a4a6e337480abfb479c94e37c5f214d"},{url:"/icons/icon-180.svg",revision:"5ecf8467a5c098136f3dc19f381ce9f5"},{url:"/icons/icon-192.png",revision:"4a4a6e337480abfb479c94e37c5f214d"},{url:"/icons/icon-192.svg",revision:"b03af01551e3a577918e83b8ca0cd1a0"},{url:"/icons/icon-512.png",revision:"4a4a6e337480abfb479c94e37c5f214d"},{url:"/icons/icon-512.svg",revision:"c8c7a3d8724e525ca96ac4eec657b237"},{url:"/manifest.json",revision:"a788175b9fcdef4648f6be0d20ecb2ed"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:c})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/.*\.tile\.openstreetmap\.org\/.*/i,new e.CacheFirst({cacheName:"openstreetmap-tiles",plugins:[new e.ExpirationPlugin({maxEntries:100,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/^https:\/\/nominatim\.openstreetmap\.org\/.*/i,new e.NetworkFirst({cacheName:"nominatim-api",plugins:[new e.ExpirationPlugin({maxEntries:50,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/^http:\/\/router\.project-osrm\.org\/.*/i,new e.NetworkFirst({cacheName:"osrm-api",plugins:[new e.ExpirationPlugin({maxEntries:50,maxAgeSeconds:3600})]}),"GET")});
+if(!self.define){let e,a={};const i=(i,s)=>(i=new URL(i+".js",s).href,a[i]||new Promise(a=>{if("document"in self){const e=document.createElement("script");e.src=i,e.onload=a,document.head.appendChild(e)}else e=i,importScripts(i),a()}).then(()=>{let e=a[i];if(!e)throw new Error(`Module ${i} didn’t register its module`);return e}));self.define=(s,t)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(a[c])return;let n={};const d=e=>i(e,c),r={module:{uri:c},exports:n,require:d};a[c]=Promise.all(s.map(e=>r[e]||d(e))).then(e=>(t(...e),n))}}define(["./workbox-f52fd911"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"afb778bed1232ec14f383e60461739f3"},{url:"/_next/static/chunks/23-e49c25ff4a2d28a9.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/294-2fa965b3689286af.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/3-2f6302d2d82ccbf4.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/51-3dd90a3c681c6cc8.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/638-fd1a0257b54ad914.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/696.643551c3285e05af.js",revision:"643551c3285e05af"},{url:"/_next/static/chunks/89-4980629c0d60c4b7.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/aaea2bcf-751e554790cca40d.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/app/_not-found/page-93d783c8c64ea450.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/app/dashboard/page-b6182a5778b047c3.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/app/demo/page-93e84db19a0bdb4b.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/app/layout-d6a93d3cfd64ad7c.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/app/page-4caceea7b779a2fb.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/d0deef33.afff9b3bf7059c8b.js",revision:"afff9b3bf7059c8b"},{url:"/_next/static/chunks/fd9d1056-a7223fb5cd24b6ec.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/framework-f66176bb897dc684.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/main-aa26475113dfe328.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/main-app-28eecd8be4b1a924.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/pages/_app-6a626577ffa902a4.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/pages/_error-1be831200e60c5c0.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-b12133d75871ed5d.js",revision:"iKCByOIh3lTAA7aty94gd"},{url:"/_next/static/css/857aa62baa17c894.css",revision:"857aa62baa17c894"},{url:"/_next/static/css/ab0f62cb2562f093.css",revision:"ab0f62cb2562f093"},{url:"/_next/static/iKCByOIh3lTAA7aty94gd/_buildManifest.js",revision:"2ec694eb52ae4f523f265a46bae4d768"},{url:"/_next/static/iKCByOIh3lTAA7aty94gd/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/media/26d4368bf94c0ec4-s.p.woff2",revision:"d28498895ed7050c37c94ddd026a4670"},{url:"/_next/static/media/2801417b65625cf5-s.woff2",revision:"d313794bd9fda8b0ad44189b27348eab"},{url:"/_next/static/media/28793f5c5e3d822d-s.woff2",revision:"c352fb162d499702d53f0dbfaab7b58f"},{url:"/_next/static/media/558ca1a6aa3cb55e-s.p.woff2",revision:"570751c5f8b418972c1976160ba6ed85"},{url:"/_next/static/media/64d784ea54a4acde-s.woff2",revision:"8a5b33d747f0cfaac631ad00bd5bcba2"},{url:"/_next/static/media/6c25f6e897d845a3-s.p.woff2",revision:"1a662f8f0b096a4b0954f9a6017ad87d"},{url:"/_next/static/media/6d831b18ae5b01dc-s.woff2",revision:"e6155c5cfacf3867c500daf0ebcba222"},{url:"/_next/static/media/7b19b489dc6743ba-s.woff2",revision:"fa052422282eb26cff400131e5af4da1"},{url:"/_next/static/media/8b44c7e6549520b2-s.woff2",revision:"7113097e6195b2704f2923f6989dd992"},{url:"/_next/static/media/ac0e76ddaeeb7981-s.woff2",revision:"6465b62dd12646a816e0d80f024ab07f"},{url:"/_next/static/media/b3bf17a9041d9433-s.woff2",revision:"ded54b0f07a0919cf7e27d7880a56e0a"},{url:"/_next/static/media/c9c3823090ec8b55-s.woff2",revision:"625be018001610bec1a5c0ba45d9988c"},{url:"/_next/static/media/edc640959b0c7826-s.woff2",revision:"5508edf7c10fe677025b8c88a2578acb"},{url:"/_next/static/media/ff71da380fbe67dd-s.woff2",revision:"60d32697500d4779da3725134067ad31"},{url:"/_next/static/media/layers-2x.9859cd12.png",revision:"9859cd12"},{url:"/_next/static/media/layers.ef6db872.png",revision:"ef6db872"},{url:"/_next/static/media/marker-icon.d577052a.png",revision:"d577052a"},{url:"/icon.svg",revision:"4a4a6e337480abfb479c94e37c5f214d"},{url:"/icons/icon-180.png",revision:"4a4a6e337480abfb479c94e37c5f214d"},{url:"/icons/icon-180.svg",revision:"5ecf8467a5c098136f3dc19f381ce9f5"},{url:"/icons/icon-192.png",revision:"4a4a6e337480abfb479c94e37c5f214d"},{url:"/icons/icon-192.svg",revision:"b03af01551e3a577918e83b8ca0cd1a0"},{url:"/icons/icon-512.png",revision:"4a4a6e337480abfb479c94e37c5f214d"},{url:"/icons/icon-512.svg",revision:"c8c7a3d8724e525ca96ac4eec657b237"},{url:"/manifest.json",revision:"a788175b9fcdef4648f6be0d20ecb2ed"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:a,event:i,state:s})=>a&&"opaqueredirect"===a.type?new Response(a.body,{status:200,statusText:"OK",headers:a.headers}):a}]}),"GET"),e.registerRoute(/^https:\/\/.*\.tile\.openstreetmap\.org\/.*/i,new e.CacheFirst({cacheName:"openstreetmap-tiles",plugins:[new e.ExpirationPlugin({maxEntries:100,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/^https:\/\/nominatim\.openstreetmap\.org\/.*/i,new e.NetworkFirst({cacheName:"nominatim-api",plugins:[new e.ExpirationPlugin({maxEntries:50,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/^http:\/\/router\.project-osrm\.org\/.*/i,new e.NetworkFirst({cacheName:"osrm-api",plugins:[new e.ExpirationPlugin({maxEntries:50,maxAgeSeconds:3600})]}),"GET")});
